Remove redundant cast

+++ src/fheroes2/maps/map_random_generator_helper.cpp
@@ -244,7 +244,7 @@ namespace Maps::Random_Generator
             const double distance = std::sqrt( static_cast<double>( dx * dx + dy * dy ) );
             const uint32_t noise = static_cast<int>( Rand::GetWithGen( 0, 2, randomGenerator ) );
             const size_t ring = static_cast<size_t>( std::floor( distance ) + noise );
-            if ( static_cast<size_t>( ring ) >= buckets.size() ) {
+            if ( ring >= buckets.size() ) {
                 buckets.resize( ring + 1 );
             }
